+++ docs/changeLog.md
@@ -2,6 +2,7 @@
 Current version: 1.1.0-beta
 
 ## 1.1.0-beta
+- Added the texture manager (TBD and shared textures)
 - Reverted the change where the input element would round the value
 - Added support for dragging entities (TBD if draggable is true)
 - Added the hoverForBanks attribute that determins whether the mouse cursor has to hover over the bank inorder to change the value